+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.142.0] - Not released
+### Added
+- Show found users and groups on the search results page.
 
 ## [1.141.3] - 2025-10-28
 ### Fixed